그냥 첨부터 다시하자 안되겠다

+++ src/main/java/com/ehdndqls/shuttle/drivers/Drivers.java
@@ -14,10 +14,10 @@ public class Drivers {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long driverId;
+    private Integer driverId;
 
     @Column(nullable = false)
-    private Long organizationId;
+    private Integer organizationId;
 
     private String driverName;
     private Integer joinYear;
