+++ src/main/java/com/ehdndqls/shuttle/dto/DriverForm.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @Setter
 public class DriverForm {
 
-    private Long id;
+    private Integer id;
     private String driverName;
     private Integer joinYear;
     private Drivers.DriverType type;
@@ -21,7 +21,7 @@ public class DriverForm {
     }
 
     // 모든 필드를 사용하는 생성자
-    public DriverForm(Long id, String driverName, Integer joinYear, Drivers.DriverType type, Boolean active) {
+    public DriverForm(Integer id, String driverName, Integer joinYear, Drivers.DriverType type, Boolean active) {
         this.id = id;
         this.driverName = driverName;
         this.joinYear = joinYear;
