+++ src/main/java/com/ehdndqls/shuttle/vehicles/VehiclesRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface VehiclesRepository extends JpaRepository<Vehicles, Long> {
+public interface VehiclesRepository extends JpaRepository<Vehicles, Integer> {
 
-    Page<Vehicles> findByOrganizationId(Long organizationId, Pageable pageable);
+    Page<Vehicles> findByOrganizationId(Integer organizationId, Pageable pageable);
 
     @Query("SELECT v FROM Vehicles v WHERE " +
             "(:searchText IS NULL OR v.vehicleModel LIKE %:searchText% OR v.vehicleNumber = :searchText) AND " +
@@ -21,7 +21,7 @@ public interface VehiclesRepository extends JpaRepository<Vehicles, Long> {
             @Param("searchText") String searchText,
             @Param("vehicleType") Vehicles.VehicleType vehicleType,
             @Param("vehicleYear") Integer vehicleYear,
-            @Param("organizationId") Long organizationId
+            @Param("organizationId") Integer organizationId
     );
 }
